requirements.txt
beautifulsoup4==4.11.2
python-dotenv==1.0.0
streamlit==1.37.0
aiohttp==3.10.2
beautifulsoup4==4.11.2
python-dotenv==1.0.0
requests
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
scikit-learn==0.24.2
